// frontend/web/src/services/API/proserve-wb-packaging-api/models/GetComponentVersionResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Workbench Packaging API
 * API in charge of packaging workbench products
 *
 * The version of the OpenAPI document: 2023-10-31T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { ComponentVersion } from './ComponentVersion';
import {
    ComponentVersionFromJSON,
    ComponentVersionFromJSONTyped,
    ComponentVersionToJSON,
} from './ComponentVersion';

/**
 * 
 * @export
 * @interface GetComponentVersionResponse
 */
export interface GetComponentVersionResponse {
    /**
     * 
     * @type {ComponentVersion}
     * @memberof GetComponentVersionResponse
     */
    componentVersion: ComponentVersion;
    /**
     * 
     * @type {object}
     * @memberof GetComponentVersionResponse
     */
    yamlDefinition?: object;
    /**
     * 
     * @type {string}
     * @memberof GetComponentVersionResponse
     */
    yamlDefinitionB64?: string;
}

/**
 * Check if a given object implements the GetComponentVersionResponse interface.
 */
export function instanceOfGetComponentVersionResponse(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "componentVersion" in value;

    return isInstance;
}

export function GetComponentVersionResponseFromJSON(json: any): GetComponentVersionResponse {
    return GetComponentVersionResponseFromJSONTyped(json, false);
}

export function GetComponentVersionResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): GetComponentVersionResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'componentVersion': ComponentVersionFromJSON(json['component_version']),
        'yamlDefinition': !exists(json, 'yaml_definition') ? undefined : json['yaml_definition'],
        'yamlDefinitionB64': !exists(json, 'yaml_definition_b64') ? undefined : json['yaml_definition_b64'],
    };
}

export function GetComponentVersionResponseToJSON(value?: GetComponentVersionResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'component_version': ComponentVersionToJSON(value.componentVersion),
        'yaml_definition': value.yamlDefinition,
        'yaml_definition_b64': value.yamlDefinitionB64,
    };
}